+++ src/test/java/com/openggf/tests/TestS3kHczSnakeBlocksObject.java
@@ -1,5 +1,6 @@
 package com.openggf.tests;
 
+import com.openggf.game.GameServices;
 import com.openggf.game.RuntimeManager;
 import com.openggf.game.sonic3k.objects.HCZSnakeBlocksObjectInstance;
 import com.openggf.game.sonic3k.objects.Sonic3kObjectRegistry;
@@ -103,7 +104,7 @@ class TestS3kHczSnakeBlocksObject {
         HCZSnakeBlocksObjectInstance instance = new HCZSnakeBlocksObjectInstance(
                 new ObjectSpawn(0x1400, 0x0600, OBJECT_ID, 0x00, 0x00, false, 0));
         instance.setServices(new DefaultObjectServices(RuntimeManager.getCurrent()));
-        RuntimeManager.getCurrent().getCamera().setX((short) 0);
+        GameServices.camera().setX((short) 0);
         instance.update(0, null);
         assertTrue(instance.isDestroyed());
     }
@@ -113,7 +114,7 @@ class TestS3kHczSnakeBlocksObject {
         HCZSnakeBlocksObjectInstance instance = new HCZSnakeBlocksObjectInstance(
                 new ObjectSpawn(0x0100, 0x0600, OBJECT_ID, 0x00, 0x00, false, 0));
         instance.setServices(new DefaultObjectServices(RuntimeManager.getCurrent()));
-        RuntimeManager.getCurrent().getCamera().setX((short) 0x1400);
+        GameServices.camera().setX((short) 0x1400);
         instance.update(0, null);
         assertTrue(instance.isDestroyed(), "Should despawn when camera is far ahead");
     }
@@ -121,7 +122,7 @@ class TestS3kHczSnakeBlocksObject {
     @Test
     void doesNotDespawnWhenCameraIsWithinCoarseBackWindow() {
         HCZSnakeBlocksObjectInstance instance = createInstance(0x00, 0x00);
-        RuntimeManager.getCurrent().getCamera().setX((short) 0x0180);
+        GameServices.camera().setX((short) 0x0180);
         instance.update(0, null);
         assertFalse(instance.isDestroyed());
     }
@@ -129,7 +130,7 @@ class TestS3kHczSnakeBlocksObject {
     @Test
     void despawnsOnceCameraMovesOneChunkPastBaseXPlus128() {
         HCZSnakeBlocksObjectInstance instance = createInstance(0x00, 0x00);
-        RuntimeManager.getCurrent().getCamera().setX((short) 0x0200);
+        GameServices.camera().setX((short) 0x0200);
         instance.update(0, null);
         assertTrue(instance.isDestroyed(),
                 "Camera coarse-back should advance to 0x0180 here, putting the base X behind the allowed window");
@@ -180,7 +181,7 @@ class TestS3kHczSnakeBlocksObject {
         HCZSnakeBlocksObjectInstance instance = new HCZSnakeBlocksObjectInstance(
                 new ObjectSpawn(BASE_X, BASE_Y, OBJECT_ID, subtype, renderFlags, false, 0));
         instance.setServices(new DefaultObjectServices(RuntimeManager.getCurrent()));
-        RuntimeManager.getCurrent().getCamera().setX((short) 0);
+        GameServices.camera().setX((short) 0);
         return instance;
     }
 }